# docs/swagger.yml
openapi: '3.0.2'
info:
  title: RecyThing
  version: '1.0'
  description: "API Documentation Capstone Project RecyThing Kelompok 2"
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /report:
    post:
      tags:
        - reporting
      summary: Create a new report for rubbish or littering
      description: Endpoint for Report rubbish or Report Littering
      operationId: createReport
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                report_type:
                  type: string
                  description: Type of the report (required)
                title:
                  type: string
                  description: Title of the report (max 100 characters)
                description:
                  type: string
                  description: Description of the report (required)
                waste_type:
                  type: string
                  description: Type of waste (required)
                waste_materials:
                  type: array
                  items:
                    type: string
                  description: List of waste materials
                latitude:
                  type: number
                  format: float
                  description: Latitude coordinate (required)
                longitude:
                  type: number
                  format: float
                  description: Longitude coordinate (required)
                address:
                  type: string
                  description: Address of the report (required)
                city:
                  type: string
                  description: City of the report (required)
                province:
                  type: string
                  description: Province of the report (required)
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: List of image files
        required: true
      responses:
        '201':
          description: Report Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: report created!
                  data:
                    $ref: '#/components/schemas/ReportResponse'
        '400':
          description: Invalid request, Upload Image Failed (filetype error, filesize to large error, dll)
      security:
        - Bearer: []

    get:
      tags:
        - reporting
      summary: Get all history user reports limit 10
      description: Endpoint for Get all reports from user history
      operationId: historyUserReports
      responses:
        '200':
          description: Get History User Reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: ok
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportResponse'
        '401':
          description: Unauthorized
        '500':
          description: Status internal error               
      security:
        - Bearer: []
...
  /user/{userId}:
    get:
      tags:
        - manage users
      summary: Get user by id
      description: Endpoint user to get user by id
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID of the user to be deleted
          example: USR0001
      security:
        - Bearer: []
      responses:
        '200':
          description: User successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: ok
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: USR0001
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "john.doe@gmail.com"
                      phone_number:
                        type: string
                        example: 089511223344
                      points:
                        type: integer
                        example: 100
                      gender:
                        type: string
                        example: laki-laki
                      birth_date:
                        type: string
                        format: date
                        example: 1990-01-01
                      address:
                        type: string
                        example: 123 Main Street, City, Country
                      profile_picture:
                        type: string
                        example: http://example.com/yourimage.png
        '404':
          description: User not found
    delete:
      tags:
        - manage users
      summary: Delete user
      description: Endpoint admin to delete a user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID of the user to be deleted
          example: U00001
      security:
        - Bearer: []
      responses:
        '200':
          description: User successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: user successfully deleted!
        '404':
          description: User not found
...
  /faqs/category:
    get:
      tags:
        - faqs
      summary: Get FAQs by category
      description: Endpoint user for get faqs by category
      operationId: getFaqsByCategory
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
            example: point dan level
          description: The name of the FAQ category to retrieve,
                       valid input (profil, littering, rubbish, misi, lokasi sampah, poin dan level, artikel)
      responses:
        '200':
          description: Get all data by category
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: ok
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FAQs'
  /faqs/search:
    get:
      tags:
        - faqs
      summary: Search FAQs by keyword
      description: Searches for FAQs that match a specific keyword
      operationId: getFaqsByKeyword
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
            example: ikon
          description: The keyword to search for FAQs.
      responses:
        '200':
          description: Get all data by keyword
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: ok
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FAQs'
...
components:
  schemas:
...
    ReportResponse:
      type: object
      properties:
        id:
          type: string
          description: ID of the created report
          example: RPT0006
        author_id:
          type: string
          description: ID of the report author
          example: USR0003
        report_type:
          type: string
          description: Type of the report
          example: rubbish
        title:
          type: string
          description: Title of the report
          example: ""
        description:
          type: string
          description: Description of the report
          example: Tempat sampah di dekat halte bus sudah penuh dengan sampah.
        waste_type:
          type: string
          description: Type of waste
          example: sampah basah,sampah kering
        latitude:
          type: number
          format: float
          description: Latitude coordinate
          example: -6.1754
        longitude:
          type: number
          format: float
          description: Longitude coordinate
          example: 106.8272
        address:
          type: string
          description: Address of the report
          example: Jl. Sudirman
        city:
          type: string
          description: City of the report
          example: Jakarta
        province:
          type: string
          description: Province of the report
          example: DKI Jakarta
        status:
          type: string
          description: Status of the report
          example: need review
        waste_materials:
          type: array
          items:
            $ref: '#/components/schemas/WasteMaterial'
        report_images:
          type: array
          items:
            type: string
            format: uri
            description: URL of the report images
          example:
            - https://res.cloudinary.com/dlbbsdd3a/image/upload/v1717008205/recything/reports/dpz5ztmx9lo65ivcqbot.jpg
            - https://res.cloudinary.com/dlbbsdd3a/image/upload/v1717008205/recything/reports/zwk2gotn8tywjewpzkhn.jpg
        created_at:
          type: string
          format: date-time
          description: Date and time when the report was created
          example: "2024-05-30T01:43:26.482+07:00"
    WasteMaterial:
      type: object
      properties:
        id:
          type: string
          description: ID of the waste material
          example: MTR04
        type:
          type: string
          description: Type of the waste material
          example: kertas
...
    FAQs:
      type: object
      properties:
        id: 
          type: string
          example: FAQ01
        category:
          type: string
          example: profil
        question:
          type: string
          example: Bagaimana cara saya memperbarui informasi profil saya?
        answer:
          type: string
          example: Anda dapat memperbarui informasi profil Anda melalui menu 'Pengaturan Profil' di aplikasi. Klik ikon profil, pilih 'Pengaturan', dan edit informasi yang diperlukan.
...